// src/proto/handleMessage.ts
export class CheckVariableResponse extends pb_1.Message {
    #one_of_decls: number[][] = [];
    constructor(data?: any[] | {
        status?: number;
        responseMessage?: string;
        message?: string;
    }) {
        super();
        pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
        if (!Array.isArray(data) && typeof data == "object") {
            if ("status" in data && data.status != undefined) {
                this.status = data.status;
            }
            if ("responseMessage" in data && data.responseMessage != undefined) {
                this.responseMessage = data.responseMessage;
            }
            if ("message" in data && data.message != undefined) {
                this.message = data.message;
            }
        }
    }
    get status() {
        return pb_1.Message.getFieldWithDefault(this, 1, 0) as number;
    }
    set status(value: number) {
        pb_1.Message.setField(this, 1, value);
    }
    get responseMessage() {
        return pb_1.Message.getFieldWithDefault(this, 2, "") as string;
    }
    set responseMessage(value: string) {
        pb_1.Message.setField(this, 2, value);
    }
    get message() {
        return pb_1.Message.getFieldWithDefault(this, 3, "") as string;
    }
    set message(value: string) {
        pb_1.Message.setField(this, 3, value);
    }
    static fromObject(data: {
        status?: number;
        responseMessage?: string;
        message?: string;
    }): CheckVariableResponse {
        const message = new CheckVariableResponse({});
        if (data.status != null) {
            message.status = data.status;
        }
        if (data.responseMessage != null) {
            message.responseMessage = data.responseMessage;
        }
        if (data.message != null) {
            message.message = data.message;
        }
        return message;
    }
    toObject() {
        const data: {
            status?: number;
            responseMessage?: string;
            message?: string;
        } = {};
        if (this.status != null) {
            data.status = this.status;
        }
        if (this.responseMessage != null) {
            data.responseMessage = this.responseMessage;
        }
        if (this.message != null) {
            data.message = this.message;
        }
        return data;
    }
    serialize(): Uint8Array;
    serialize(w: pb_1.BinaryWriter): void;
    serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
        const writer = w || new pb_1.BinaryWriter();
        if (this.status != 0)
            writer.writeUint32(1, this.status);
        if (this.responseMessage.length)
            writer.writeString(2, this.responseMessage);
        if (this.message.length)
            writer.writeString(3, this.message);
        if (!w)
            return writer.getResultBuffer();
    }
    static deserialize(bytes: Uint8Array | pb_1.BinaryReader): CheckVariableResponse {
        const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new CheckVariableResponse();
        while (reader.nextField()) {
            if (reader.isEndGroup())
                break;
            switch (reader.getFieldNumber()) {
                case 1:
                    message.status = reader.readUint32();
                    break;
                case 2:
                    message.responseMessage = reader.readString();
                    break;
                case 3:
                    message.message = reader.readString();
                    break;
                default: reader.skipField();
            }
        }
        return message;
    }
    serializeBinary(): Uint8Array {
        return this.serialize();
    }
    static deserializeBinary(bytes: Uint8Array): CheckVariableResponse {
        return CheckVariableResponse.deserialize(bytes);
    }
}
interface GrpcUnaryServiceInterface<P, R> {
    (message: P, metadata: grpc_1.Metadata, options: grpc_1.CallOptions, callback: grpc_1.requestCallback<R>): grpc_1.ClientUnaryCall;
    (message: P, metadata: grpc_1.Metadata, callback: grpc_1.requestCallback<R>): grpc_1.ClientUnaryCall;
    (message: P, options: grpc_1.CallOptions, callback: grpc_1.requestCallback<R>): grpc_1.ClientUnaryCall;
    (message: P, callback: grpc_1.requestCallback<R>): grpc_1.ClientUnaryCall;
}
// ...
export abstract class UnimplementedvariableServiceService {
    static definition = {
        checkVariable: {
            path: "/variableService/checkVariable",
            requestStream: false,
            responseStream: false,
            requestSerialize: (message: CheckVariableRequest) => Buffer.from(message.serialize()),
            requestDeserialize: (bytes: Buffer) => CheckVariableRequest.deserialize(new Uint8Array(bytes)),
            responseSerialize: (message: CheckVariableResponse) => Buffer.from(message.serialize()),
            responseDeserialize: (bytes: Buffer) => CheckVariableResponse.deserialize(new Uint8Array(bytes))
        }
    };
    [method: string]: grpc_1.UntypedHandleCall;
    abstract checkVariable(call: grpc_1.ServerUnaryCall<CheckVariableRequest, CheckVariableResponse>, callback: grpc_1.sendUnaryData<CheckVariableResponse>): void;
}
export class variableServiceClient extends grpc_1.makeGenericClientConstructor(UnimplementedvariableServiceService.definition, "variableService", {}) {
    constructor(address: string, credentials: grpc_1.ChannelCredentials, options?: Partial<grpc_1.ChannelOptions>) {
        super(address, credentials, options);
    }
    checkVariable: GrpcUnaryServiceInterface<CheckVariableRequest, CheckVariableResponse> = (message: CheckVariableRequest, metadata: grpc_1.Metadata | grpc_1.CallOptions | grpc_1.requestCallback<CheckVariableResponse>, options?: grpc_1.CallOptions | grpc_1.requestCallback<CheckVariableResponse>, callback?: grpc_1.requestCallback<CheckVariableResponse>): grpc_1.ClientUnaryCall => {
        return super.checkVariable(message, metadata, options, callback);
    };
}